add session helper
